Fix tests and move shared code to a command helper
