+++ app/commands/decidim/solutions/admin/create_solution.rb
@@ -6,6 +6,7 @@ module Decidim
       # A command with all the business logic when a user creates a new solution.
       class CreateSolution < Decidim::Command
         include Decidim::Challenges
+        include Decidim::CommandUtils
 
         # Public: Initializes the command.
         #
